+++ public/index.html
@@ -8,6 +8,6 @@
     </head>
     <body>
         <div id="root"></div>
-        <script src="./src/index.pack.js"></script>
+        
     </body>
 </html>
